refact common selectable list items code
